+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- show total number of unfinished todos in navigation badge
+
 ## [0.7.0] - 2024-03-29
 
 - do not count finished todos in sidebar badges